Add new methods to NoOpCookieStore
rg/http4s/client/asynchttpclient/NoopCookieStore.scala
