+++ async-http-client/src/main/scala/org/http4s/client/asynchttpclient/AsyncHttpClient.scala
@@ -20,9 +20,6 @@ import org.http4s.internal.invokeCallback
 import org.http4s.util.threads._
 import org.log4s.getLogger
 import org.reactivestreams.Publisher
-import _root_.io.netty.handler.codec.http.cookie.Cookie
-import org.asynchttpclient.uri.Uri
-import org.asynchttpclient.cookie.CookieStore
 
 object AsyncHttpClient {
   private[this] val logger = getLogger
@@ -156,12 +153,4 @@ object AsyncHttpClient {
     Headers(headers.asScala.map { header =>
       Header(header.getKey, header.getValue)
     }.toList)
-  private class NoOpCookieStore extends CookieStore {
-    val empty: java.util.List[Cookie] = new java.util.ArrayList()
-    override def add(uri: Uri, cookie: Cookie): Unit = ()
-    override def get(uri: Uri): java.util.List[Cookie] = empty
-    override def getAll(): java.util.List[Cookie] = empty
-    override def remove(pred: java.util.function.Predicate[Cookie]): Boolean = false
-    override def clear(): Boolean = false
-  }
 }
